move Accessor logic into Configuration
